refactor(ui): move stories to tests/storybook
